+++ app/components/DraggableComponent.ts
@@ -1,28 +1,18 @@
 import React, { useState } from 'react'
-import Draggable, { DraggableEventHandler, ControlPosition, DraggableEvent } from 'react-draggable'
+import Draggable, { DraggableEventHandler, DraggableProps } from 'react-draggable'
 import { useSelector } from 'react-redux'
-import { StateWithHistory } from '../util/defaultState'
+import { State } from '../util/defaultState'
 
-type PositionDataHandler = (data: ControlPosition) => void | false
-
-type DraggableComponentProps = {
-  children: React.ReactNode,
-  handle: string,
-  position?: ControlPosition,
-  onStop: PositionDataHandler,
-  onDrag?: PositionDataHandler,
-  onStart?: DraggableEventHandler,
-  onClick?: DraggableEventHandler,
-  disabled?: boolean,
-  enableUserSelectHack?: boolean
+interface DraggableComponentProps extends DraggableProps {
+  onClick?: DraggableEventHandler
 }
 
-const ZEROZERO = { x: 0, y: 0 }
-
-// Wrapper around Draggable that can also handle click events
+// Wrapper around Draggable from react-draggable which
+//   - adds another callback, onClick, for click events
+//   - set the scale = state.display.svgZoom
 export default function DraggableComponent(props: DraggableComponentProps) {
-  const [dragStartPos, setDragStartPos] = useState(ZEROZERO)
-  const svgZoom = useSelector<StateWithHistory, number>(state=> state.display.svgZoom)
+  const [dragStartPos, setDragStartPos] = useState({ x: 0, y: 0})
+  const scale = useSelector<State, number>(state=> state.display.svgZoom)
 
   const onStart: DraggableEventHandler = (evt, data) => {
     // @ts-ignore
@@ -31,29 +21,21 @@ export default function DraggableComponent(props: DraggableComponentProps) {
   }
 
   const onDrag: DraggableEventHandler = (evt, data) => {
-    props.onDrag && props.onDrag({ x: data.x, y: data.y })
+    props.onDrag && props.onDrag(evt, data)
   }
 
   // see https://github.com/react-grid-layout/react-draggable/issues/531
   const onStop: DraggableEventHandler = (evt, data) => {
     // calls onClick instead of onStop if mouse has not moved (or barely moved)
+    // @ts-ignore
     if (Math.abs(dragStartPos.x - evt.screenX) < 5 || Math.abs(dragStartPos.y - evt.screenY) < 5 ) {
       props.onClick && props.onClick(evt, data)
     } else {
-      props.onStop && props.onStop({ x: data.x, y: data.y })
+      props.onStop && props.onStop(evt, data)
     }
   }
 
-  const draggableProps = {
-    onDrag, onStop, onStart,
-    scale: svgZoom,
-    handle: props.handle,
-    disabled: Boolean(props.disabled),
-    // Setting the position to 0,0 has the effect of ensuring that all drag deltas always start with 0,0.
-    // The onStop and onDrag callbacks all work off of relative coordinates.
-    // position: props.position || ZEROZERO
-    // enableUserSelectHack: props.enableUserSelectHack
-  }
+  const draggableProps = Object.assign({}, props, { onDrag, onStop, onStart, scale })
 
   return React.createElement(Draggable, draggableProps, props.children)
 }
