+++ app/components/Node.tsx
@@ -1,92 +1,74 @@
 import React, { useCallback } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
+import partial from 'lodash/partial'
 import DraggableComponent from './DraggableComponent'
+import { DraggableEventHandler } from 'react-draggable'
+import NodeBody from './NodeBody'
 import NodeHalo from './NodeHalo'
 import NodeCircle from './NodeCircle'
 import NodeImage from './NodeImage'
 import NodeLabel from './NodeLabel'
-import ConditionalLink from './ConditionalLink'
-import { StateWithHistory } from '../util/defaultState'
 import type { Node as NodeType } from '../graph/node'
-import type { ControlPosition, DraggableEvent } from  'react-draggable'
+import type { State } from '../util/defaultState'
+import { getNodeUIState, NodeUIState } from '../util/NodeUIState'
 
-type NodeProps = {
-  id: string,
-  currentlyEdited: boolean,
-  selected: boolean,
-  status: string
-}
+// This is our main node component
+// <Node>
+//   <DraggableComponent>
+//     <NodeBody> (<g>, root in dom)
+//       <NodeHalo>
+//       <NodeLabel>
+//       <NodeCircle>
+//       <NodeImage>
 
-export function Node({ id, currentlyEdited, selected, status } : NodeProps) {
+export default function Node({ id }: { id: string }) {
   const dispatch = useDispatch()
-  const node = useSelector<StateWithHistory, NodeType>(state => state.graph.nodes[id])
-  const editMode = useSelector<StateWithHistory>(state => state.display.modes.editor)
-  const isSelecting = useSelector<StateWithHistory, boolean>(state => state.display.selection.isSelecting)
-  const isHighlighting = useSelector<StateWithHistory, boolean>(state => state.annotations.isHighlighting)
-  const selectedNodeIds = useSelector<StateWithHistory, string[]>(state => state.display.selection.node)
-  const isMultipleSelection = selected && selectedNodeIds.length > 1
-  const showSelection = Boolean(editMode && (selected || currentlyEdited))
+  const editMode = useSelector<State>(state => state.display.modes.editor)
+  const node = useSelector<State, NodeType>(state => state.graph.nodes[id])
+  const uiState = useSelector<State, NodeUIState>(partial(getNodeUIState, id))
 
-  const moveNode = useCallback( (deltas: ControlPosition) => {
-    if (isMultipleSelection) {
-      return false
-    } else {
-      dispatch({ type: 'RELEASE_NODE', id, deltas })
-    }
-  }, [dispatch, id, isMultipleSelection])
+  const onStart = () => {
+    dispatch({ type: 'DRAG_NODE_STOP', id })
+  }
 
-  // the id in the payload, while otherwise redundant, allows redux-undo
-  // to group drag actions into a single action
-  const dragNode = useCallback( (deltas: ControlPosition)  => {
-    if (isMultipleSelection) {
-      return false
-    } else {
-      dispatch({ type: 'DRAG_NODE', id, node, deltas })
-    }
-  }, [dispatch, id, node, isMultipleSelection])
+  const onStop: DraggableEventHandler = (_event, data) => {
+    const deltas = { x: data.x, y: data.y }
+    dispatch({ type: 'DRAG_NODE_STOP', id, deltas })
+  }
 
-  const clickNode = useCallback( (event: DraggableEvent) => {
-    if (!editMode && node.url) {
-      // Draggable prevents ordinary click events coming from its drag handle elements,
-      // so we have to open the link programmatically
-      window.open(node.url, "_blank")
-    } else if (isSelecting) {
-      dispatch({ type: 'SWAP_NODE_SELECTION', id })
-    } else if (isHighlighting) {
-      dispatch({ type: 'SWAP_NODE_HIGHLIGHT', id })
-    } else {
-      dispatch({ type: 'CLICK_NODE', id })
-    }
-  }, [dispatch, isSelecting, isHighlighting, id, editMode, node])
+  const onDrag: DraggableEventHandler = (_event, data) => {
+    const deltas = { x: data.x, y: data.y }
+    dispatch({ type: 'DRAG_NODE', id, deltas })
+  }
 
-  const onMouseEnter = useCallback(() => {
-    dispatch({ type: 'MOUSE_ENTERED_NODE', name: node.name })
-  }, [dispatch, node.name])
+  const onClickDraggable = () => {
+    dispatch({ type: 'CLICK_NODE', id })
+  }
 
-  const onMouseLeave = useCallback(() => {
-    dispatch({ type: 'MOUSE_LEFT_NODE' })
-  }, [dispatch])
+  const onMouseEnter = () => {
+    dispatch({ type: 'MOUSE_ENTERED_NODE', id })
+  }
 
-
-  // to prevent HTML5 drag-n-drop (draggable="false" used to work)
-  //const onDragStart = (e: DraggableEvent) => e.preventDefault()
+  const onMouseLeave = () => {
+    dispatch({ type: 'MOUSE_LEFT_NODE', id })
+  }
 
   return (
-    <DraggableComponent handle=".draggable-node-handle" onStop={moveNode} onClick={clickNode} onDrag={dragNode}>
-      <g id={"node-" + id} className="oligrapher-node" onMouseEnter={onMouseEnter} onMouseLeave={onMouseLeave} onDragOver={onMouseEnter} onDragLeave={onMouseLeave}>
-        <ConditionalLink condition={Boolean(!editMode && node.url)} url={node.url}>
-          { node.description && <title>{node.description}</title> }
-          <ConditionalLink condition={Boolean(editMode && node.url)} url={node.url}>
-            <NodeLabel node={node} status={status} />
-          </ConditionalLink>
-          <NodeHalo node={node} selected={showSelection} highlighted={status === "highlighted"} />
-          <NodeCircle node={node} status={status} />
-          <NodeImage node={node} status={status} />
-        </ConditionalLink>
-      </g>
+    <DraggableComponent disable={editMode} handle=".draggable-node-handle" onStart={onStart} onStop={onStop} onDrag={onDrag} onClick={onClickDraggable}>
+      <NodeBody nodeId={node.id} url={node.url} enableClick={!editMode} onMouseEnter={onMouseEnter} onMouseLeave={onMouseLeave}>
+        <NodeLabel node={node} uiState={uiState} />
+        <NodeHalo node={node} uiState={uiState} />
+        <NodeCircle node={node} uiState={uiState} />
+        { node.image && <NodeImage node={node} uiState={uiState} /> }
+      </NodeBody>
     </DraggableComponent>
   )
 
 }
 
-export default React.memo(Node)
+// // the id in the payload, while otherwise redundant, allows redux-undo to group drag actions into a single action
+  // to prevent HTML5 drag-n-drop (draggable="false" used to work)
+  //const onDragStart = (e: DraggableEvent) => e.preventDefault()
+  // <DraggableComponent handle=".draggable-node-handle" onStop={onStop} onClick={onClick} onDrag={onDrag}>
+  // onDragLeave={onMouseLeave}
+  // { node.description && <title>{node.description}</title>}
